swift: deploy swift_ring_manager to one node per cluster
This is part of a number of changes related to deploying new swift
ring management automation.

This CR defines a ring_manager_host per production cluster, and
deploys the swift_ring_manager software there, alongside the
appropriate YAML config file for that cluster, a systemd timer to run
the manager every hour, and an rsync server to make the new rings
available to the puppetmasters. These things are removed from any host
that is not the ring_manager_host, to reduce the risk of having more
than one manager process running on the same cluster.

Deploying this CR alone will not change any production rings (there's
a separate CR to arrange for updated rings to be collected on a
puppetmaster; and one to subsequently deploy those updated rings on
swift nodes).

This is a variant approach to hieradata management, where we define a
common hiera structure with swift cluster metadata in (so it's
available from both puppetmasters and swift/thanos nodes), and move
towards defining cluster_label which is used to look up that cluster
metadata. Here we only do the minimal changes necessary to make it
work for ring management (i.e. essentially only frontends), it could
be extended further later.

Part of this involves moving swift::stats_reporter from profile to
class (and making use of statsd_{host,port} already being available in
the two profiles that make use of this class).

Tested in pontoon.

Hosts: ms-fe2009.codfw.wmnet,ms-fe1009.eqiad.wmnet,thanos-fe1001.eqiad.wmnet
Bug: T265117
Change-Id: I6bc2f8fa117a161a1b90c768f4e0dedc0c3a55c5