alertmanager: add basic wmcs routing rules
This adds a basic configuration for routing team: wmcs alerts to
cloud-feed on IRC. This can be adjusted later, but it should give us a
good start for defining Prometheus based alerts.

Someone with ChanServ founder access to the #wikimedia-cloud-feed (I am
not one of those people) should grant the jinxer-wm bot autovoice before
merging this using this command:
 /msg ChanServ FLAGS #wikimedia-cloud-feed jinxer-wm +V
That way the bot doesn't get tripped on Libera.Chat antispam features
as easily.

Bug: T302493
Change-Id: Icc0d576f4e434b11ef45834668fcdbbbb0a6dd4e