Add several ASNs to those that alert as critical from Icinga
Our BGP checks against network elements normally alert at 'warning'
level in Icinga.  A list of ASNs is passed to the script which should
instead throw a 'critical' level alert if peers on them are down.

The patch adds a few new ASNs to the list:

1. AS5511 - Orange, now transit provider in drmrs
2. AS14907 - WMF public ASN.  L3 switches peer to CRs using this AS.
3. AS64810 - Used for EVPN iBGP peerings between switches in new
eqiad rows e/f.

Bug: T299758
Change-Id: I08f6df54ef0063454830278402a02bacf96c1807