gerrit: prevent 'null' entry in email
Sync up the Comment template with upstream source at v3.3.9. It removes
the variable ${group.link} which is no more set by Gerrit and cause
email to output a 'null' string.

Update the comment block to refer to the current Gerrit version, I have
verified we are in sync with upstream.

Add a reference to T43608 which is the reason we forked the template.

Once deployed by Puppet, Gerrit will automatically detect the template
update thus no further action is required.

Bug: T288312
Change-Id: I0cc73ce48a46ebff951a207d7c88df03072d07d3