puppet: Print nodes that change on every puppet run, sorted
Alerts for node(s) that change on every puppet run are printed
in random order every time.

This causes 2 minor inconveniences- changing the message of the
icinga text with its state not changing can cause additional
alerts; and the text messages can vary from node to node.

The addition of a sorted() should order the node names in
lexicographical order, making reading the alerts much easier.

Change-Id: I2672866faee86afe8d9ac59f54efeb472266805c