Move some common resources to the opensearch::server profile
The opensearch::server profile is a generic profile for using opensearch,
whereas opensearch::server::logstash is the specific implementation for
logstash.

We are adding a second specific implementation, so it makes sense to
move any shared resources from the logstash to the generic profile.

This includes a plugin directory and the prometheus exporter per
instance. This change should be a no-op for existing logstash nodes.

Bug: T301382
Change-Id: I003e1d7208d7beb93eab4a604b78792b1d1e02a6