puppetmaster: rsync swift rings from each cluster's ring manager
This is part of a number of changes related to deploying new swift
ring management automation.

This change makes the puppetmaster master frontend rsync the new rings
tarball from the swift ring manager host into volatile, so it can then
be deployed out across the cluster (that change is in a different CR).

Deploying this CR will not change any production rings (there's
another CR to actually deploy the new rings to the clusters)

Tested in pontoon

Hosts: re:puppetmaster.*wmnet
Bug: T265117
Change-Id: I71202c9db4775e21ccdf4a7c25f51f670d3286ab