prometheus: Add more per-index metrics for elasticsearch
We want to create phabricator tickets when specific indices get
older than a cutoff. To make that possible start storing the
index age in prometheus, so alertmanager can be configured to
create tickets for old indices.

While we only need to check two specific wikis for this task,
those wikis are already covered by the current set of indices
being collected. Collect the index age metric for the same
set of indices we collect against normally.

Bug: T300295
Change-Id: Ia233a8489e0a761b4719e299eedb9d3ebd8c1e07