Add role::insetup for ms-fe1012
The node has being racked/reimaged recently, and it is currently
showing networking problems. To avoid the catch-all definition
above it, add a more specific one to site.pp.

The other nodes of the batch, ms-fe10[09-11] are already working
with the prod role (so remove the role::insetup definition that
is not applied).

Bug: T294137
Change-Id: Ib0e47a3a26a5607821291a5936cbf375619fcf05