Stop pinning the TGC cookie to the user agent and IP address
By default the TGC is bound to the IP address from which it was created
(and a static user agent), the former causes session expiries if people
are roaming with their laptops (or in some case by an ISP which hands
out ipv4 and ipv6 addresses variantly).

This was tested in idp-test.wikimedia.org, so also enable it in
production.

Bug: T273858
Change-Id: I51fe64d46aa5d2893f87bc5c7366c72d5cfeb229