varnish: allow injecting rate-limiting rules from hiera
This is a stopgap solution to allow us to insert rate-limiting
rules from the private puppet repository  while we set up a proper
dynamic rate-limiting solution.

This is a huge footgun of course, but we consider its immediate utility
to outweigh the risks, if we're going to eventually remove it in a short
time.

Bug: T303534
Change-Id: I8aae51bc3303ccb05c05b88379cef875264e0513