certspotter: add -start_at_end to only fetch new logs
As per the documentation,

    Start monitoring logs from the end, rather than the beginning.  This
    significantly reduces the time to run Cert Spotter, but you will miss
    certificates that were added to a log before Cert Spotter started
    monitoring it.

This is a possible fix for T303593 but more investigation might be
required. In general, the purpose of restarting the certspotter service
is so that we can be aware of the possible new certificates that are
issued; we do not care about the older certificates at this stage. This
will also help reduce the logs that are fetched as certspotter was
disabled in November 2020 and that means fetching a lot of data.

We don't enable the systemd timer yet: let's see how this plays out and
then we can enable that in another commit.

Bug: T303593
Hosts: O:alerting_host
Change-Id: I2d6f7abe0f73b3fb42049503078ba0eab2cf6969