Set simpler partman recipe for kubernetes200[5,6]
We have reused the old kubernetes recipe, but we don't really
need any LVM config since we have a single (virtual) 20G disk to partition.
This change adds a more standardized partition scheme for the
two VMs that already moved to bullseye + overlay (we need to reimage
them again sadly).

Bug: T300744
Change-Id: I5606a791828e03998b112557699fa4f630976e6d