backy2: move sqlite db to /srv
When reimaging cloudvirts it's relatively easy to preserve
the contents of /srv; by moving the db onto /srv
we can carry on where we left off with old backups
after a reimage.

For existing hosts I will move the db file manually at the same
time that this patch is merged.

Change-Id: I6ef7f951212f1ffa66a0a9c190e70d45abd7569d