install_server: add the flat-noswap.cfg recipe/override
In some environments we need to skip the swap partition,
so this change adds an override to flat.cfg called
flat-noswap.cfg. While testing this change me and Filippo
realized that there was a bug in flat.cfg as well,
namely a '..' preventing the expert_recipe to be applied
by partman without errors (partman was silently defaulting
to its base settings, very close to what flat.cfg is
meant to do). This change also fixes the long standing bug
in flat.cfg.

Change-Id: Ibf24583cab0f2e13a494c10071a03c479acd7183