bacula: Add mixed priority to all jobs
From the manual:

When set to yes (default no), this job may run even if lower
priority jobs are already running. This means a high priority job
will not have to wait for other jobs to finish before starting.
The scheduler will only mix priorities when all running jobs have
this set to true.

This should solve some of the scheduling issues when restoring or
performing multiple backups.

Bug: T95705
Change-Id: I6b4d98409e871f20ac128fd14fc3354fd433c99b