dnsdist: remove redundant rate limits
We had some rate limits in place to prevent misuse of the service when
we initially started working on it. This commit removes those limits and
we will make use of dynamic blocks in its place when needed.

We do not remove the configuration option entirely but instead make it
configurable, with a default setting of disabled.

Hosts: O:wikidough
Change-Id: I92bdf50641318614fab188762ca68d94ce05015a