role::kafka::logging: add PKI migration settings
The settings are needed to allow upgrading one broker at the time
with a new PKI-based TLS certificate.

The brokers trust each others' TLS certs using a set of
TLS CN values, that for the moment is only one
(CN=kafka_logging-eqiad_broker).
We have to add to the list of trusted admins all the
hostnames of the brokers, since the new PKI-based certs will
have as CN the broker's hostname (as expected, to allow
proper hostname verification etc..).

Last but not the least, the broker's truststore will need to
change as happened for clients, to trust boh Puppet and PKI
TLS certs.

This change will require a roll restart of the brokers.

Bug: T300130
Change-Id: I45a237131948f5e207bdfed13d320d6bc54a8087