P:icinga: add profile for performance tweaking
In T204993, we re-enabled certspotter on the alerting host to track the
issuance of certificates. Since it has been enabled, there has been an
increase in the network activity [0] on the host as certspotter fetches
logs from multiple certificate transparency log servers, resulting in
bursty traffic when it runs every 30 minutes. Relatedly, there also has
been an increase in the network errors (rx_discards) that happen at the
same time certspotter runs.

We believe this is due to the bursty nature of the traffic and that the
Icinga host drops the packets because of that. Since we don't want to
reduce the number of CT log servers certspotter fetches the logs from
(to reduce the network activity), we should improve the network
performance of the host itself.

This is the first commit in possibly a series of commits to try to
alleviate this issue. We first start by setting interface::rps to the
alerting_host role in order to improve the network performance. In the
next commit, we will try to adjust the RX queuelen, which is currently
set to 200 (with a maximum support of 2047).

[0] - see the alert1001 Grafana dashboard

Hosts: P:icinga
Bug: T303593
Change-Id: I6f29e955bf3563a98a02f2c4da4942230f2cc512