karapace: use karapace included python; set hostname
The original karapace deb unpacked itself into the root of the debian
filesystem, which overwrote the system python. This caused prometheus to
fail and would have caused other issues as well.

The new one installs itself to /usr/lib/karapace/bin, so the systemd
unit needed an updated path.

Additionally I was seeing the following error in the karapace systemd unit:

    karapace1001 karapace[1489148]: ValueError: Can't build URL with "port" but without "host".

Although it is documented as allowing an empty string here:
https://github.com/aiven/karapace#configuration-keys

this error which went away when the host was added to the karapace config.

Bug: T301565
Hosts: karapace1001.eqiad.wmnet
Change-Id: I4798ee78bf4b04c7729b1d33fb325aaaed220798