elastic: fix cirrus settings check false negative
This patch fixes a bug where the lists were sorted
in-place, whereas the script was expecting the sort function
to actually return the value of the new sorted objects.

The result of the bug was a false negative.

--

This reverts commit 69f73d1e6311a4a4dab2d199aeec371f9d9cba62.

Reason for revert: Fixed underlying metrics issue in
https://phabricator.wikimedia.org/T301511#7798267

Bug: T301511
Hosts: elastic1054.eqiad.wmnet, elastic2054.codfw.wmnet
Change-Id: I4c13637ad195d24f130ca78bbeefb2ab4dfe886b