elastic: fix cirrus settings check false negative
This patch fixes a bug where the lists were sorted
in-place, whereas the script was expecting the sort function
to actually return the value of the new sorted objects.

The result of the bug was a false negative.

--

This reverts commit ae60707326e29fe04dd64e14ac7dfff07b7cf460.

Reason for revert: Fixed underlying metrics issue in
https://phabricator.wikimedia.org/T301511#7798267

Bug: T301511
Hosts: elastic1054.eqiad.wmnet, elastic2054.codfw.wmnet
Change-Id: Ic325720ab8886132696735585d231bdcce2bfe3a