swift: Create a new read-only role on mw account for backup taking
In order to minimize the impact that a potential bug or security
breach could have on the media backups service, create a separate
role for the mediawiki account, with separate credentials and
limited access rights: read only, which will be what's needed
to download and backup files (no need for writes or admin
privileges).

This will require a separate and previous deploy to the private
puppet repo adding a new key for this role.

Bug: T269108
Change-Id: I0422e0f6dfba45886b2ab148a0b61d704bab4eb4