dnsdist: remove redundant rate limits (attempt 2)
This is a continuation of commit 435af904d3: since AddAction() is a
dynamic rule and can be added without restarting dnsdist, let's
completely remove it from the config file and run it from the console
when desired. We don't want default limits for the same reasons as
described in the earlier commit.

Hosts: O:wikidough
Change-Id: I4b2bb4d524c8eb8b2f49410aa0d87ee49bf9973f