gitlab: add version check to restore script
This change introduces a version check before the restore.
In the past a manual step was needed to deactivate
and activate restore of replica while doing GitLab updates. It
would be good if the restore is just skipped when a version mismatch
occurs, without stop services and puppet on the replica.
Especially if we start doing backup and restore more frequent.

Bug: T274463
Bug: T285867
Change-Id: Icbe7e2bac9c3e7991a5acf782746674fa673f36a