Move cloudstore1008/1009 to role::spare
The NFS workloads previously handled here (maps and scratch)
have been moved to ceph_VM servers. I'd like to keep
these files around for a month or two as a safety net,
but the first step in knowing we don't need these servers
is turning off the NFS servers which this change will allow.

Bug: T291405
Change-Id: I2a3344c659d69bc59e27d2f4c518657867745d76