hieradata: generate cfssl certs for cloudmetrics*
This also adds grafana-cloud.wm.o to the SANs. That name is already
present in DNS and the caches, but hasn't been working previously
because the old-style SSL cert did not have that as a SAN. Migrating to
the new cfssl cert system instead of regenerating the old-style certs
with the new name is easier.

Change-Id: Ia176f156aea66b6dadb6c0d845bbe623158faebb