geoip::maxmind: rename the update timers, don't use 'legacy' term
Instead of using the term "legacy" use the term "main" since
the term is overloaded.

To differentiate this from the other timer, which can't have the same
name of just geoip_update, we are calling this "main" and the
other "ipinfo" because those databases are needed for the
IPInfo extension.

Absent old timer and job, then clean up code afterwards.

Bug: T303464
Change-Id: I0398d167f120faffb88e170cc87475d6956a00a7