Stop using profile::base::linux419 on Hadoop nodes
This was originally applied to unbundle the Hadoop update to Buster (first
the cluster was migrated to 4.19 while still on Stretch and then the OS
was upgraded). This migration is long complete and all of the Hadoop cluster
is based on Buster, so remove the old includes and Hiera settings.

Hosts: C:profile::base::linux419
Change-Id: I55b690db2de20333a78062578ac30c108f61f32e