ci: on castor server drop /srv requirement
The castor server requires extended disk space to be mounted on `/srv`
which was expressed by Mount['/srv'] requirement.

When the role is applied to an instance with a Cinder volume attached,
it already mounted and not managed by Puppet and we have to set:

    profile::ci::slave::labs::common::manage_srv: false

This prevent the LVM mount from happening and does not create a
Mount['/srv'] resource.

Remove the require from castor server
Ensure the dependency between profile in the role.

Bug: T252071
Change-Id: I45440fbdcc549484e373b221af88435a43171714