gitlab: remove conditional, always use gitlab::restore class
Remove the conditional around the instantiation of the gitlab::restore
class.

The parameters passed to it are already set to either present or absent
and that should be sufficient to make sure the files get either
installed or removed.

If we additionally do this and not even use the class in some cases
then it would never actively remove the files once the parameters
are changed to 'absent'.

And it leads to confusion when trying to compile other changes
that are suddenly "noop" like on this one: Icbe7e2bac9c3e799.

So we should always use the class, one way or another, but not just
skip it.

Bug: T274463
Change-Id: I7518265864639af0fcb40436353aecf58d14e6c3