mariadb: Use ROW binlog_format for db_inventory.
Orchestrator's queries are not safe using STATEMENT replication, causing
mariadb to issue this warning:

```
[Warning] Unsafe statement written to the binary log using statement
format since BINLOG_FORMAT = STATEMENT. REPLACE... SELECT is unsafe
because the order in which rows are retrieved by the SELECT determines
which (if any) rows are replaced. This order cannot be predicted and may
differ on master and the slave
```

I've already made the change to the live systems.

Bug: T301315
Change-Id: I9d0c3bb7beab47c63e84774970f35dcd40fe80ab