logstash: populate target index format and add pipeline diagnostics
Adds a feature to write a `logstash` hash with diagnostic data if
the ENABLE_DIAGNOSTICS environment variable is set to "true".  It is
disabled by default but could later be enabled for CI to test if an
event will go to the intended index pattern.

Populates fields required for generating the target index.
The template revision field is intentionally excluded because it is a
Puppet concern.

This change should have no effect on how logs are rendered and indexed.

Bug: T305090
Bug: T305175
Change-Id: I5245e3689a7e88cf91e361204ee8a31852073b14