zotero: Disable paging
Why:
Zotero has been paging on and off every now and then, but starting
2022-03-28 it has been paging every day, sometimes multiple times a day.
There is literally no action that SREs can take aside from a restart. As
pointed out in T291707#7810937, this is not sustainable and is
demoralizing long term.

This is due to a defficiency in zotero that does not allow a non
functioning instance to be depooled automatically by kubernetes
currently. It that is to be addressed in
https://gerrit.wikimedia.org/r/774848 and some followup helm chart
changes that will be required. But until that happens, and due to a slew
of new pages today (2022-04-02) it's best to stop paging.

What:
Switch page attribute to false. SREs will stop receiving notifications
on their mobile devices. IRC notification will continue as before

Bug: T291707
Change-Id: I4dd83eb7cbb61b1dabcf7f27a2ee91daaf9294f8