haproxy: use Requires= in haproxy-mtail@tls.service
As described in T305275, when the cp host is restarted, the
haproxy-mtail@tls.socket fails with the following message:

    haproxy-mtail@tls.socket: Socket service haproxy-mtail@tls.service
    already active, refusing.

This happens because the haproxy-mtail@tls.service currently specifies
"After=" for the haproxy-mtail@tls.socket; additionally, specifying
"Require=" will ensure that the socket is always ready before the
service is started.  (We cannot pass a socket to an existing service
without restarting the service, so the socket should be started *before*
the service.)

Why does the solution 'systemctl stop haproxy-mtail@tls.service;
systemctl stop haproxy-mtail@tls.socket; puppet agent -t' as presented
in T305275 work?  Presumably because the order of Puppet execution
activates the socket first and then the service, but we have to specify
the same order and the dependency for systemd independently, i.e., start
the socket first and then the service.

Bug: T305275
Hosts: O:cache::upload_haproxy
Change-Id: I529dc4bb5d4b93b8a05de2cb85bd272f91fec3fa