trafficserver: update Icinga check in check_trafficserver_log_fifo.py
In T305275, we identified an issue with the reboot of the cp hosts where
the reboot-single cookbook fails to complete execution because of a
failed Icinga check:

    Not all services are recovered:
    cp2036:check_trafficserver_log_fifo_notpurge_backend

This happens because in check_trafficserver_log_fifo.py, the script
exits with status UNKNOWN when it detects the service is not pooled
which is true because the cookbook script depools the service before it
starts, however, the UNKNOWN status instead of OK prevents the cookbook
from completing execution.

This commit returns "OK: not pooled" so that the cookbook can finish
executing and pool the service automatically on completion.

Bug: T305275
Hosts: O:cache::upload_proxy
Change-Id: I7c76d6337e7814f467f0c86b6ec130d0df11e34e