cloudgw1001: use a custom name for the dataplane NIC
The predictable NIC name used by systemd is too long (was enp101s0f1np1).
When adding the vlan tags, the interface name was simply invalid (> 16 which is
IFNAMSIZ).

The cloudgw code does a lot of {name}.{vlan} by default, so instead of changing all
the puppet code to support the 'vlanXXXX' scheme, simply override the interface name
with a systemd .link file.

Additional information:
 https://manpages.debian.org/bullseye/udev/systemd.link.5.en.html#EXAMPLES

Bug: T304598
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: Id661989a8d908ceccff0cdcb28f6104261deb873