external_clouds_vendors: Support entity types besides "cloud"
For now, cloud ipblocks are still the only ones represented, but this
refactor makes it possible to add other types, for example to get the IP
space used by Googlebot or other well-known clients.

The name "cloud" is still used in a bunch of places, including the
script filename; we can adjust those in a followup to reduce confusion.

Bug: T305581
Change-Id: I58eb50205337faf74de362a319f0931d4326d56e