varnish: Expand static.php optimisation regarless of query string
The /w/ static requests used to only be deterministic if a 5-char
hex was passed as query string, which /w/static.php did (and still
does) use to serve the right file from any of the available MW dirs.

As of this week, the handling for URLs without such hash is now
deterministic as well (based on the newest MW branch, instead of
the MW branch for the "current" wiki). The responses no longer vary
by wiki ID or $IP in wmf-config, so this optimisation can be expanded
and thus also enforcing that static.php can't regress in the future
since it will always see the same wiki after this patch.

Bug: T302465
Change-Id: I65f86214990b9695e2fa420b9b3f95c78434a079