mediawiki: Update httpbb tests for /static/current going away
Update the positive test to verify /w/static.php instead, which was
strangely missing from both of these.

Enable the negative test, which is passing on the /w/ version already
on both prod and k8s. The /static/current path will be a 404 soon
more generally, which is probably not worth testing for right now.

Bug: T302465
Change-Id: I3cd083bcadfa75da40e733b1b6f62b66ac46a2da