phabricator: allow disabling ssh-phab service except on one host
Parameterize the state of the ssh-phab (git-ssh.wikimedia.org) service
instead of hardcoding it.

Make the default 'absent' because we are going to deprecate it.

But for now leave it running on phab1001 and phab2001
while this allows us to apply the phabricator role on new hosts
without starting this.

Bug: T296022
Change-Id: Iee531bd369971fb35679329171b261253698dee1