openstack: remove horizon access to puppetmaster
Horizon only needed access to the puppetmaster api for the 'guided mode'
which was removed even before I became active around here[0]. This
commit removes the Horizon settings and firewall rules used for that,
since it does not seem we will be re-introducing that feature in the
future.

Puppet needs and will need access to the ENC api. Those are not changed
by this commit.

[0]: I believe this was because of a Puppet upgrade moving the API
     endpoints it needed to an enterprise-only feature. :(

Change-Id: Id5c12821851caf5a4dae44e24745105443eb926b