Increase threshold for IPv6 RIPE Atlas probes from 65 to 90
We have seen a bunch of alerts for RIPE IPv6 probes to Eqiad today.
After extensive checking no sign of a specific problem on our side,
or a specific pattern in terms of transit provider, peer, or source
AS the probes are coming from, has emerged.

The overall level of failure has increased by 2-3% only, but we were
relatively close to the configured alert threshold for this alert
before, and given that increase the check is firing noise frequently.

This change increases the threshold to stop the noise for now, netops
will review the situation again next week and decide what the best
longer-term value for this is, and also review if the situation has
changed.

Bug: T305703
Change-Id: Iab464293ce1d6973e4bc4f15e6f7f92596b936c5