gitlab_runner: solve race condition to to make things work on first run
When creating a new gitlab-runner and applying the puppet
role for the first time there are some dependency issues.

It tries to work on config files before those exist.

Things are generally fine after the second puppet run
but this is an attempt to make it a little bit nicer
by working on first run.

Change-Id: Ic49473bc4eb38f7161614deba7f957564c72b043