mx: reject email to legacy mailing list domains
Prior to this change we forwarded legacy mailing list traffic to
lists.wikimedia.org.

After this change we reject mail to <list-name>@wikimedia.org and
<list-name>@wikipedia.org with a 551 error code indicating the correct
address. This patch does not reject list names with suffixes such as
-bounce or -owner. Doing so, though possible, would have made the
configuration much more complex.  Given that most of the mail to the
legacy domains is spam, this config should be sufficient.

This change is a step towards moving wikimedia.org's MX records
to gmail.

Hosts: O:mail::mx
Bug: T280472
Change-Id: Ic5046dfe5cc9094241b5c01e57c1eaef24b16b6a