DHCP: make doh and durum hosts use the bullseye installer
Regardless if you end up using a cookbook or not
you will need this to tell DHCP next install should be on bullseye.

That is if you want to keep the host names, but you said you do,
so here it is.

Can be merged anytime unless you'd be worried that they get booted
into PXE by accident before you do the reimages but I don't see that
as a a real risk.

Bug: T305589
Change-Id: I1e133414339bc964089805e02712612c2c81d6f5