Use both dbproxy101[89] servers for both wikireplica services
To-date we have used a single dbproxy server for each of the
wikireplicas-a and wikireplicas-b services. Whilst this works it does
complicate scheduled maintenance.

This commit adds the other dbproxy server to the service catalog, so
that they are both in use at all times. When required, they could each
be depooled with confctl.

Hosts: lvs1019.eqiad.wmnet,lvs1020.eqiad.wmnet
Hosts: dbproxy1018.eqiad.wmnet,dbproxy1019.eqiad.wmnet
Bug: T298940
Bug: T298586
Change-Id: I5fc66841ca96b7a0525d5ddbbed9e06450a08bbd