Don't prompt for loading additional firmware in d-i
hw-detect in Bullseye asks if unavailable firmware should be loaded from
a USB drive (or floppy). In general we add all available firmware to our
internal d-i boot images, but there's the case of the tg3 driver which
can optionally load firmware (but we use it without).

As such preseed the dialogue to not prompt in general for bullseye installs.

Bug: T306148
Change-Id: Ie6d4871fb0d6f046401c756cbdd356a73dfd5d45