admin: Add placeholder to reserve uid and gid 914 for minio-user
minio contains hundreds of millions of files as an offline backup
for MediaWiki multimedia files. Changing the uid of its files
takes over a week. Reserve uid and gid 914 for it, in order to
avoid changing it every time a host reimage while preserving the
host data happens.

This will not be applied until all files are migrated to the new
id (currently they use uid/gtid 499- except backup2007, which
was painfully changed to 498).

Bug: T305446
Change-Id: Ie66c2250f199f7869cd3ce4c232943309705b322