gitlab: temp set gitlab-runner user to root for bootstrapping gitlab-runner2001
Originally, and for non protected runners, the gitlab-runner user is
root.

Then in Ief404a77968120e77caa it was changed to a non-privileged user.

When applying the role on a new machine / after reimaging and running
puppet for the first time the non-root user can't create the home
dir and register due to permissions.

PANIC: mkdir /home/gitlab-runner/.gitlab-runner: permission denied etc

So assuming the normal procedure to create runners is to temp change
to root, run puppet once and then change it back to non-privileged user.

Bug: T297659
Change-Id: I426d9bd6958867e8a6d1b14e01fd57dd33946e0f