mediabackup: Preconfigure mc client config on worker nodes
mc (minio client) uses json to provide easy access to a predefined
set of MinioIO servers. While the application doesn't require the
client to work, it is quite convenient to have it automatically
added to the root user, so it is available for an administrator to
check or modify its data out of band.

This was manually configured for my user only, puppetize it for
root and it can be soft-linked for other users on demand.

Bug: T305446
Change-Id: Ib7c2706b676f5e0069a6a47e15ea645c1d7b742d