P:wikidough: add a check to ensure service has been restarted
This commits adds an nrpe::monitor_service check to make sure that if a
configuration file was changed, that the related service was restarted
within a given time interval.

For a more practical example (and the reason for this commit): some
services such as dnsdist and pdns-recursor (both used by Wikidough)
don't support reloading of the configuration and so their systemd
service has to be restarted whenever a configuration file changes. But
restarting also means clearing out the cache, so ideally we would like
to manage such restarts ourselves versus letting Puppet do that and
because of that, any change to a configuration file should not result in
an automatic restart.

The purpose of this script is to check if a configuration file change
(mtime) is more recent than the subsequent service restart
(ActiveEnterTimestamp), to remind us to restart the service (WARNING).
If the time delta between the mtime and ActiveEnterTimestamp exceeds 24
hours (configurable), then it raises the alert from WARNING to CRITICAL.

Hosts: O:wikidough
Change-Id: Ibe9390bcee4b68115f68ffa45e8feb73a5351025