gitlab: ensure home dir for runner_user exists when running as non-root
When trying to apply the puppet role on a new host for the first time
and running as non-root user there are errors because the non-privileged
user can't create the home directory.

But if you switch to root then it's not created. Ensure it's created
when using root to be able to bootstrap a host as root, run puppet once
and then switch to non-root.

Bug: T297659
Change-Id: Ia5838c5cf1f72c74fab6d965967cd514a0a8a57a