gitlab_runner: ensure the full path to the config location exists
follow-up to Ia5838c5cf1f72c74fab

We need to make sure not only /home/gitlab-runner/ (or whatever user
name we use) exists, but also /home/gitlab-runner/.gitlab-runner.

During registration the config is written to this place after
reading from the config _template_ file under /etc/gitlab-runner
and without the hidden directory inside the home dir that fails
with more permission issues for the non-privileged user.

Since we use mkdir -p we just have to give it the latter to have both.

Bug: T297659
Change-Id: I31515df83f97b0cafa57b4e899a79081a8d32e10