docker: ensure apparmor package is installed if on bullseye
On bullseye apparmor is now installed by default as opposed to buster.
This is the "libapparmor1" package.

The userspace utilities package "apparmor" is not installed by default
though.

Both come from the 'apparmor' source package but are not pulling in each
other.

This leads to Debian bug #989781 and you can't run containers.

The docker-gc and docker-resource-monitor services will also fail.

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=989781

After installing the apparmor package this is resolved.

This happened to me when upgrading gitlab-runners to bullseye.

For now I did the fix in the gitlab-runner class but it seems
like this should be in the docker class in general.

Change-Id: I2c914347d31233154842d1a41513eac9759fa779