gitlab::runner: if on buster, ensure apparmor package is installed
On bullseye apparmor is now installed by default as opposed to buster.
This is the "libapparmor1" package.

The userspace utilities package "apparmor" is not installed by default
though.

Both come from the 'apparmor' source package but are not pulling in each
other.

This leads to Debian bug #989781 and you can't run containers.

The docker-gc and docker-resource-monitor services will also fail.

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=989781

After installing the apparmor package this is resolved.

This happened to me when upgrading gitlab-runners to bullseye.

Fixing this here for now but also made I2c914347d312331548 to
do it for any host using docker on bullseye, separately.

Bug: T297659
Change-Id: I503eb1ede927182e171c123a4b952a6d3c9f4981