P:wikidough: update notes_url to point to new monitoring page
We have pointed notes_url for monitoring::service to the Wikidough
landing page for quite a while now, even though there was no monitoring
information present there.

It's time to update the notes to their respective sections on the new
monitoring page so that we can properly reference the respective Icinga
checks: what they are and how they can be resolved in case there is a
failure (and an alert on IRC).

Hosts: O:wikidough
Change-Id: I9b97f7102746bbe92ad524205e91027dd7e79d2f