swift: wmf/rewrite.py say 400 earlier if passed bad UTF-8
We see clients sending us url-encoded invalid UTF-8. Catch that early,
and just say 400 bad request straight away (rather than letting the
UnicodeDecodeError be caught by the handler in handle_request which
logs a backtrace) - we only want to be logging a backtrace if we've
had a problem with processing valid input, rather than if we were sent
garbage in the first place.

Bug: T305942
Change-Id: I1b92501d3fe58b59e7986988fc8c1a7acd8fd7b3