gerrit: keep computing changes mergeability
Gerrit 3.4 defaults to no more compute the mergeability of open changes
when the target branch has been updated. The reason is it is an O(N^2)
indexing operations which can be slow on large repositories. We had it
for a while and beside the spikes of reindexing tasks, it does not seem
to cause any trouble.

Explicitly enable the mergeability (the default in 3.3) so we don't
loose it when upgrading to Gerrit 3.4

Bug: T303970
Change-Id: I73476306d9169207920cbf7061e63e5f73c6cb3c