install_server/gitlab: separate partman recipes for physical servers
So far all gitlab* machines have been VMs.

But now we just got physical hardware in both eqiad (T301177)
and codfw T301183.

This means we can't use the wildcard anymore and need to separate
partman config between VMs and physical servers by number.
(Or we would have to use different hostnames for both types).

The tickets asked for RAID 1 already and confirmed with dcops
there are 2 disks in them, so using the raid1-2dev.cfg here.

This will unblock OS installation of the new hosts in both DCs.

Bug: T301183
Bug: T301177
Change-Id: I11f2ea3a01f9178ed97318adf1e2b98d2977ca6f