prometheus::blackbox::check: add new blackbox exporter check
This Cr adds a new resource for managing blackbox-exporter checks.  it
makes use of exported resources to all uses to create defines in the
application profiles that run on the host.

based of the work in the following patchset
 https://gerrit.wikimedia.org/r/c/operations/puppet/+/786384/3

This is a PoC to foster some comments.  its based of the work used to
generate the service::catalog probs but could definitely be missing
something

Hosts: P:prometheus::ops
Change-Id: Iac7c7aecd0687ac0ad8218f2b069c3c433150096