lvm::volume: add createonly flag and use in cinder backups
Turns out that the device names can change with kernel changes (at
least!) and that will try to zap/etc. devices that it should not try
to. This introduces a flag that will avoid trying to do anything to
the devices as long as the volume group already exists, see:

https://forge.puppet.com/modules/puppetlabs/lvm#optional-values

Bug: T307117
Change-Id: Idfe994e086429777b091d21a19515212f9f93cdf
Signed-off-by: David Caro <dcaro@wikimedia.org>