openstack: update tools-redis to a 'new' style name
As a part of migrating tools-redis to bullseye and automatic failover
I'm changing the canonical service name to the 'modern' style (meaning
it's in svc.$PROJECT.eqiad1.wikimedia.cloud.). Update the legacy names
(managed by wmcs-wikireplica-dns) to point to the new name so that
there's only one actual name to update when failing over to the new
redis cluster.

Bug: T278541
Change-Id: I7abb652f21616433b9ccc9cfb93970a56d4f5ddf