Add new profile to build the CAS debs
We used to build these on the main build hosts in the past, but that has
a few downsides:
- The OS of the build host may differ from the CAS setup and the CAS debs
  are build for the current OS only (no pbuilder)
- The build hosts added Java 8 to their package lists, which breaks the
  Gradle build if present (it detects Java 8 and attempts to use it,
  while it needs 11)

Instead create a small dedicated build directory and some rsync config
to allow to pull the deb for import to apt.wikimedia.org. This gets applied
to the idp-test hosts which are close to the IDP install in production.

Hosts: idp-test1001.wikimedia.org
Change-Id: I2602e8e7f9496ad3c1efdaacc23721f11e07fb39