P:wikidough: do not automatically restart the pdns service
With 144f4ff482 merged, we should not restart the pdns-recursor service
when the configuration file changes as we want to manually control the
restart of the service.

As the service restarts will be manual after this commit is merged,
f517d8026c is a related change that notifies us in case we changed the
configuration file but forgot to restart the service.

Hosts: O:wikidough
Change-Id: Ib01973876654a581a7ad22fffed1c1bc01c64cfe