Expand stick-table test to three other hosts
Per the last few weeks of performance data on cp2027, there seems to be
minimal overhead in tracking these stats.  A small increase in CPU usage,
and potentially a slight increase in memory used:
CPU usage: https://w.wiki/58Cq
memory: https://w.wiki/58Cr

So I suggest expanding this data collection to three hosts in busier sites.
The overhead will be higher there, as traffic is higher, but it seems very
unlikely to be a problem.


Bug: T306580
Change-Id: I676e349bce5c268b111e2055c4188cec8030e254